feat : change image res form & manage res form by seperate files
이미지 response url 배열로 url 정보 전달
db의 select 객체와 api res 형식을 responses, mappers 폴더의 js 파일을 이용해 관리

참고 : responses와 mappers로 무리하게 관리하는 부분은 삭제 예정
