+++ src/responses/user-res.js
@@ -1,0 +1,8 @@
+export const userSelect = {
+  id: true,
+  email: true,
+  name: true,
+  nickname: true,
+  image: true,
+  createdAt: true,
+};
